refactor: numeric_cast -> numeric::cast
