+++ docs/tasks/2021y-03m-12d-0061-numeric.md
@@ -1,0 +1,29 @@
+
+|     дата начала     |     дата конца      | длительность    | исполнитель  |
+|:-------------------:|:-------------------:|:---------------:|:------------:|
+| 2020y-03m-13d 19:40 | 2020y-03m-13d 23:25 | 3 hours, 45 min | Kartonagnick |
+
+numeric
+----
+
+Рефактор numeric_cast
+
+1. `numeric_cast.hpp` переезжает в `tools/numeric.hpp`  
+2. вся функциональность переезжает в `namespace numeric`  
+3. из имен функций удаляем слово `numeric`
+
+было:
+```
+tools::numeric_cast<ret>(from)
+```
+
+стало:
+```
+tools::numeric::cast<ret>(from)
+```
+
+План:  
+ - [ ] рефактор кода.  
+ - [ ] рефактор тестов.  
+ - [ ] проверяем старые компиляторы  
+ - [ ] проверяем новые компиляторы
