+++ include/tools/confbuild.hpp
@@ -17,9 +17,10 @@
 // [2021y-03m-11d][00:30:56] Idrisov Denis R. 018
 // [2021y-03m-11d][23:19:26] Idrisov Denis R. 019
 // [2021y-03m-12d][23:28:45] Idrisov Denis R. 020
+// [2021y-03m-13d][21:34:16] Idrisov Denis R. 021 PRE
 #pragma once
 #ifndef dTOOLS_CONFBUILD_USED_
-#define dTOOLS_CONFBUILD_USED_ 020
+#define dTOOLS_CONFBUILD_USED_ 021 PRE
 //==============================================================================
 //==============================================================================
 
@@ -59,7 +60,7 @@
 
 #define dVERSION_MAJOR      0
 #define dVERSION_MINOR      2
-#define dVERSION_PATCH      0
+#define dVERSION_PATCH      1
 
 #define dSTRINGIZE(...) #__VA_ARGS__
 #define dSSTRINGIZE(x) dSTRINGIZE(x)
